chore: Update UsersList component to useFilteredRightHolders hook and fix search functionality
